initial commit
This is a demo of the bug I reported as Issue 19 for Hammer.js
(https://github.com/EightMedia/hammer.js/issues/19).  There is a
one-line change (line 145) added to the hammer.js file to demonstrate
(in the console) that the hammer is not processing events on buttons
after the first bucket throw.